[fix] apiの修正と配置ディレクトリの変更
