+++ app/src/app/api/score/all/route.ts
