+++ app/src/app/api/score/week/route.ts
@@ -1,6 +1,6 @@
+import type { RankingScores } from "@/types";
 import { prisma } from "@lib/prisma";
 import { getDay, subDays } from "date-fns";
-import { type RankingScores } from "@/types";
 
 const getLastSunday = (date: Date) => {
 	const dayOfWeek = getDay(date);
@@ -11,16 +11,10 @@ export async function GET() {
 	const today = new Date();
 	const lastSunday = getLastSunday(today);
 
-	const startOfDay = new Date(lastSunday.setHours(0, 0, 0, 0));
-	const endOfDay = new Date(today.setHours(23, 59, 59, 999));
+	const startOfWeek = new Date(lastSunday.setHours(0, 0, 0, 0));
+	const endOfWeek = new Date(today.setHours(23, 59, 59, 999));
 
-    const userScores = await prisma.user.findMany({
-        where: {
-			createdAt: {
-				gte: startOfDay,
-				lte: endOfDay,
-			},
-		},
+	const userScores = await prisma.user.findMany({
 		select: {
 			id: true,
 			name: true,
@@ -29,24 +23,35 @@ export async function GET() {
 			scores: {
 				select: {
 					point: true,
+					createdAt: true,
+				},
+				where: {
+					createdAt: {
+						gte: startOfWeek,
+						lte: endOfWeek,
+					},
 				},
 			},
 		},
-		distinct: ['uid'],
+		distinct: ["uid"],
 	});
 
-	const scoreDetails: RankingScores[] = userScores.map((userScore) => {
-		const sumScore = userScore.scores.reduce((sum, score) => sum + score.point, 0);
+	const scoreDetails: RankingScores[] = userScores
+		.map((userScore) => {
+			const sumScore = userScore.scores.reduce(
+				(sum, score) => sum + score.point,
+				0,
+			);
 
-		const rankingScore: RankingScores = {
-			id: userScore.id,
-			userName: userScore?.name || "",
-			imageUrl: userScore.photoUrl,
-			totalPoint: sumScore,
-		};
-		return rankingScore;
-	}).sort((a, b) => b.totalPoint - a.totalPoint)
-
+			const rankingScore: RankingScores = {
+				id: userScore.id,
+				userName: userScore.name || "",
+				imageUrl: userScore.photoUrl,
+				totalPoint: sumScore,
+			};
+			return rankingScore;
+		})
+		.sort((a, b) => b.totalPoint - a.totalPoint);
 
 	return new Response(JSON.stringify(scoreDetails), {
 		status: 200,
